Move the featurecell to its own file
